+++ cp2k_output_tools/blocks/__init__.py
@@ -1,8 +1,9 @@
 from .condition_number import match_overlap_matrix_condition_number
+from .energies import match_energies
+from .forces import match_forces
+from .kv_sections import match_kv_sections
 from .mulliken import match_mulliken_population_analysis
 from .program_info import match_program_info
-from .kv_sections import match_kv_sections
-from .energies import match_energies
 from .warnings import match_warnings
 
 builtin_matchers = [
@@ -11,5 +12,6 @@ builtin_matchers = [
     match_program_info,
     match_kv_sections,
     match_energies,
+    match_forces,
     match_warnings,
 ]