+++ tests/test_mulliken.py
@@ -1,13 +1,13 @@
 import pytest
 
-from . import TEST_DIR
-
-from cp2k_output_tools.parser import parse_iter
 from cp2k_output_tools.blocks import match_mulliken_population_analysis
+from cp2k_output_tools.parser import parse_iter
+
+from . import TEST_DIR
 
 
 def test_mulliken_unrestricted():
-    with open(TEST_DIR.joinpath("inputs/mulliken_unrestricted_snippet.out"), "r") as fhandle:
+    with open(TEST_DIR.joinpath("outputs/mulliken_unrestricted_snippet.out"), "r") as fhandle:
         result = next(parse_iter(fhandle.read(), matchers=[match_mulliken_population_analysis]))
 
         assert result
@@ -37,7 +37,7 @@ def test_mulliken_unrestricted():
 
 
 def test_mulliken_restricted():
-    with open(TEST_DIR.joinpath("inputs/mulliken_restricted_snippet.out"), "r") as fhandle:
+    with open(TEST_DIR.joinpath("outputs/mulliken_restricted_snippet.out"), "r") as fhandle:
         result = next(parse_iter(fhandle.read(), matchers=[match_mulliken_population_analysis]))
 
         assert result