+++ tests/test_warnings.py
@@ -1,11 +1,11 @@
-from . import TEST_DIR
-
-from cp2k_output_tools.parser import parse_iter
 from cp2k_output_tools.blocks import match_warnings
+from cp2k_output_tools.parser import parse_iter
+
+from . import TEST_DIR
 
 
 def test_warnings():
-    with open(TEST_DIR.joinpath("inputs/warning-snippet.out"), "r") as fhandle:
+    with open(TEST_DIR.joinpath("outputs/warning-snippet.out"), "r") as fhandle:
         result = next(parse_iter(fhandle.read(), matchers=[match_warnings]))
 
         assert result